ash: Fix crash in overview for theme change
In overview, we manage the layer opacity of the application window for
dark/light theme change to ensure the expected UI. However, for some
code paths, `overview_item_` might be null if it's about to restore. The
fix is to add a null pointer check.

Followings are repro steps without waiting for a theme change. However,
whenever there is a theme change after `OnOverviewItemWindowRestoring`
for `OverviewItemView`, the crash would happen.

Steps to repro:
  1. Open multiple application windows.
  2. Go to overview.
  3. Use one finger to drag an overview item and do not release.
  4. Use another finger to open status try.
  5. Drag up the overview item to close while quickly click on `Dark
     theme` button with another finger.

Bug: b:270171802
Test: Manual && All/DesksTest.DragWindowToCloseWithThemeChange/*
Culprit: crrev.com/c/4220145
Change-Id: I73c34c7cd100e85107c55c696e809895948f2f8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4277703
Reviewed-by: Sammie Quon <sammiequon@chromium.org>
Commit-Queue: Yongshun Liu <yongshun@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1108102}