Move in-product help trigger from PopupRowView to PopupViewViews.
The only relation that the in-product help suggestions have to an
individual row is that they are tied to a specific suggestion.
These, however, can also be handled in the parent class - there is
no need to push it down to the row level.

Bug: 1411172
Change-Id: Ic92def3b1375e7b52b9a0e540f74f458e975db5c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4274089
Commit-Queue: Jan Keitel <jkeitel@google.com>
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1108240}