[TTFCC] Don't hide TTF when Manage payment settings clicked
There is a request to continue showing the bottom sheet when navigating to settings from it and then returning baack:
https://docs.google.com/presentation/d/1Yenn4ynCYL8xlSuFdTGriu9_r5zLyulLHtUoABaDjnU/edit?resourcekey=0-DiHS6JPGwIfS1vIgpBxqqA#slide=id.g498b6e6f17907937_0

Note that this change will also not hide the bottom sheet in case when switching from the Chrome app to another android app (this was agreed).

The is aligned with the fast checkout bottom sheet.

Bug: 1247698
Change-Id: I97067eac53e4db89e5f8f4d86111ce1b6391776e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4272487
Reviewed-by: Christoph Schwering <schwering@google.com>
Reviewed-by: Viktor Semeniuk <vsemeniuk@google.com>
Commit-Queue: Anna Tsvirchkova <atsvirchkova@google.com>
Cr-Commit-Position: refs/heads/main@{#1108287}