[iOS] Refactor usage of UIButton.contentEdgeInsets Pt 1
UIButton.contentEdgeInsets is deprecated as of iOS 15. As part of the
effort of bumping the min version of iOS to 15, we refactor the usage
of UIButton.contentEdgeInsets and replace it with
UIButton.configuration.contentInsets. A particular issue in creating
this CL was that documentation gives the idea that contentEdgeInsets,
titleEdgeInsets, and imageEdgeInsets should be replaced by
contentInsets. We are asking the CL reviewer to verify that this is
correct. Additionally, we use compile guards instead of @available
because we get deprecation warnings with @available. This may be because
@available is checked at runtime and these warnings happen during compiling.

Bug: 1406561
Change-Id: I1669c371a3026b6b3ed6be133f35032f7d24e82c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4276801
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Commit-Queue: Joemer Ramos <joemerramos@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1108302}