[iOS] Add Cross Merchant Support to Ranking and Unsubscription
Before this CL, when the user navigated to a page that sold the same
product that the user was already tracking through a different site,
that product wouldn't be listed first in the Tracked Items section. In
addition, if the user unsubscribed from that item, it would not appear
in the trackable section.

In this CL, given this situation, the cross merchant product will appear
first in the tracked items section and will appear in the trackable
section on canceling the subscription. In addition, this CL fixes the
issue with comparing two products. Before this CL, two products were
identified as the same if their URLs were equal. However, there are
cases where two products could be the same despite having different
URLs. In this CL, instead of comparing products via their URLs, they are
now compared via their `product_cluster_id`, which is provided by
ShoppingService.


Bug: 1416582
Change-Id: Idf27ba05e5e85b8b3b7d281b5fa7467f03c23340
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4258365
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Ali Juma <ajuma@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1108319}