media: Add CdmEvent::kHardwareContextReset
HardwareContextReset is supposed to happens during suspend/resume or
GPU swap, which should NOT be treated as an error. Therefore, it's not
included in the hardware secure decryption fallback logic.

However, we suspect that HardwareContextReset could happen without any
suspend/resume or GPU swap. One of the symptoms is that
HardwareContextReset happens on the first session creation or generate
request, or without any playback.

This CL makes the change that HardwareContextReset without any power
or display change will be reported to MediaFoundationServiceMonitor
and be included in the hardware secure decryption fallback logic.
This is controlled under a feature param which is default to true.
We can disable this feature if the assumption is wrong and we ended up
disabling a lot of users. We may also run an experiment to see how it
works.

Manually tested by following these steps:
1.  Launch chrome with
    `--enable-features=HardwareSecureDecryptionFallback:on_hardware_context_reset/true`
1a. Force HardwareContextReset in MediaFoundationCdm when updating
    the session, and after a few playback attempts, hardware secure
    decryption is disabled due to the fallback logic.
1b. Sleep the device during playback and notice a playback error with
    DRM_E_TEE_INVALID_HWDRM_STATE. This is ignored since it's after
    a power change and therefore does not disable hardware secure
    decryption.
2.  Launch chrome with
    `--enable-features=HardwareSecureDecryptionFallback:on_hardware_context_reset/false`
    Repeat 1a and 1b and hardware secure decryption is never disabled.

Bug: 1370844
Test: Updated unit tests. See above for manual tests done.
Change-Id: Iabd4a4516ad9e66e5346a8fc3c762764b7ede1c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4015369
Auto-Submit: Xiaohan Wang <xhwang@chromium.org>
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Reviewed-by: Emily Stark <estark@chromium.org>
Reviewed-by: Frank Li <frankli@microsoft.com>
Reviewed-by: John Rummell <jrummell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1108360}