Notes: make page and tab context menu entry points work.
Reuse and update UserNotesUI for the UserNotesController which was used
in the previous views implementation as a way to pass updates from the
context menus into c/b/ui/views. Disable the context menus from the NTP
where we will not be letting users take notes. Use mojo to add a way to
let the user notes webui know, both during creation or while already
showing, that focus should move to the persistent entry note to start
the note creation process when triggered from the context menus. We
also delete a test file from the BUILD file that will break due to these
changes and tests code that is never run and will be deleted as part of
the cleanup from the previous implementation.

Bug: 1394044
Change-Id: I71b637f56b984b819fb2652a31594a8c87782f00
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4220072
Reviewed-by: Yuheng Huang <yuhengh@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: Caroline Rising <corising@chromium.org>
Reviewed-by: Brandon Wylie <wylieb@chromium.org>
Reviewed-by: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1108372}