[IBAN filling] Render full IBAN value on suggestion clicking.
This CL fixes an issue where we would autofill the masked IBAN
value when the user clicked on an IBAN suggestion, instead of
the correct unmasked value. After this CL, we autofill the full
IBAN value.
To do this, we introduce a new autofill::Suggestion payload type,
AutofillValue, that contains the unmasked IBAN value. A new payload
type was used rather than a GUID because full IBAN value can be applied
when construct suggestion instead of looking up for IBAN by using GUID.

Bug: 1344590
Change-Id: Id8dc9a6170c6b3818d26af26eeb2a96154a6866e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4248138
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Reviewed-by: Vinny Persky <vinnypersky@google.com>
Commit-Queue: Qihui Zhao <qihuizhao@google.com>
Reviewed-by: Florian Leimgruber <fleimgruber@google.com>
Cr-Commit-Position: refs/heads/main@{#1108395}