[WebEngine] Communicate when browser process fails to initialize.
Also handle any exceptions that can be caused as a result of that. This
can happen if the WebView provider version is out of date for example.

This uses a Toast to display an error message, but it is only for the
development shell application.

Change-Id: I6036da78d390a03faba6fd11e3ffde59e9044c37
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4281498
Commit-Queue: Rayan Kanso <rayankans@chromium.org>
Reviewed-by: Susanne Westphal <swestphal@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1108402}