[Tab Model] Find nearby rather than adjacent tab when closing
When closing a group of tabs in incognito mode it was possible to end
up in the normal tab switcher even if the incognito switcher still
contained tabs.

This was the result of when closing multiple tabs the TabModel would
only check for an adjacent tab that wasn't closing. When closing
multiple tabs this is insufficient and the search should look for
a nearby non closing tab.

Fix this by making the search more exhaustive, but to avoid causing
slowdown skip this behavior when closing all tabs. For multiple
tab closures, this process should only happen once when the first
tab is closed at which point a non-closing tab will be selected.

Bug: 1405321
Change-Id: I13986ee73778756e4808478f1e972665582cbf38
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4261135
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1108412}