Replace mock by FakeChromeUserManager in OfflineSigninLimiter test
The test originally used MockUserManager, but upcoming change will
break the set up so we need to fix it.
As a preparation, this replaces mock by the fake, which is maintained
in a central place, so that we can reduce the one-by-one test updates.

Along with the replacement, in FakeChromeUserManager,
- creation of the account with saml is supported.
- if LocalState is not created, fallback to take a look at global
  instance.

BUG=1325210
TEST=Ran tryjob

Change-Id: I0c56f68289693328d11dcb2b079956f95671928d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4264072
Reviewed-by: Kush Sinha <sinhak@chromium.org>
Reviewed-by: Denis Kuznetsov <antrim@chromium.org>
Reviewed-by: Roman Sorokin <rsorokin@google.com>
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1108420}