Cryptohome recovery: fix infinite loop issue when fetch rart failed
We ask the user to login again when there is no reauth proof token
present before recovery. However, it might happen that the token
is missing because there's a failure when fetching reauth request
token. In that case, the user will be stuck in an infinite loop of
logging in.
We need a flag to indicate whether the token is missing due to
fetching error and we'll allow the user to fall back to the manual
recovery option in that case.

Bug: b:257073746
Change-Id: I41a4e84458fedbdd65c24f19ca4c5ec905a562e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4282457
Commit-Queue: Yunke Zhou <yunkez@google.com>
Reviewed-by: Denis Kuznetsov <antrim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1108422}