[Thumbnail Cache] Eagerly prune cache when not tabs not visible
Make TabContentManager#updateVisibleIds() a reliable signal for which
thumbnails are needed in the cache. Prune any entries not required by
the current Android Layout. Layout and its subclasses are the primary
users of updateVisibleIds.

The primary tab id is considered to have a live layer and not require
a thumbnail in the cache. All other requested tab ids will be added
to the cache up to the configured maximum size. Any tab ids still in
the cache that are not required will be pruned when the
kThumbnailCacheRefactor flag is enabled.

This CL also does a bit of tidying up of callers of updateVisibleIds
in particular StaticLayout which now releases its visible tab ids upon
a live layer being available.

Bug: 1402843
Change-Id: Ibb82db18d2410214a1cc504f41a8a4fe8212525a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4226062
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1108437}