Records handlers which failed to consume back press.
Refactor BackPressHandler so that it can return a value to indicate
whether back press is correctly consumed. If a handler fails,
the back press will pass to remaining handlers.
Introduce BackPressResult to represent the result of handling back
press. The reason why boolean is not used is because I want to
differentiate between handleBackPress and traditional onBackPressed,
and also to emphasize not handling back press is a failure which
may have already made chrome working inappropriately.

For some handlers it might be hard to verify, so we return UNKNOWN.

This CL also adds a new histogram to record them.

Bug: 1414606, 1414832
Change-Id: Ib841f9626fc30b73682f0f8d079a65288a894850
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4251777
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Donn Denman <donnd@chromium.org>
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Reviewed-by: Sinan Sahin <sinansahin@google.com>
Commit-Queue: Lijin Shen <lazzzis@google.com>
Cr-Commit-Position: refs/heads/main@{#1108448}