[sync] Init bridge for ProxyModelTypeControllerDelegate for tests
As observed in NonUiSyncableServiceBasedModelTypeController, delegate
provider cannot be called during tests when `syncable_service` is null.
This needs to be fixed in order to allow calling
ClearMetadataWhileStopped in these cases. This CL fixes this by
initializing a fake bridge whose only purpose is to pass a delegate
instance.

Bug: 897628
Change-Id: I4611854d5397ff30904d2ae33a033ee8d0d2fa93
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4281577
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Ankush Singh <ankushkush@google.com>
Cr-Commit-Position: refs/heads/main@{#1108455}