[Read Anything] Distill with rules-based until ScreenAI is installed.
Observe ScreenAI install changes in the browser. When ScreenAI is
installed, inform the WebUI that Screen2x is ready. The WebUI
informs AXTreeDistiller which creates the mojo pipe to ScreenAI in
the utility process.

The AXTreeDistiller distills with Screen2x if the mojo pipe is bound.
If not, it distills with the rules-based algorithm.

The flow of information is:
[browser] ReadAnythingController --> ReadAnythingModel -->
ReadAnythingPageHandler
[mojo] read_anything.mojom
[renderer] ReadAnythingAppController --> AXTreeDistiller

Bug: 1266555
Change-Id: Ibc96479659490387987547ee3e8d978ed8c6ea71
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4262827
Commit-Queue: Abigail Klein <abigailbklein@google.com>
Reviewed-by: Ramin Halavati <rhalavati@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1108463}