Create mojo transport test for untrusted processes.
This CL adds a new set of tests to verify that a Transport
to an untrusted process cannot transmit file handles
that are opened writeable that are backed by executable files.

This CL also adds the `is_remote_process_untrusted` parameter
to the Transport::Create helper function.

BUG=1418101

Change-Id: I872e6c99776693dcc5b518e5538fd7277e968479
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4277627
Commit-Queue: Will Harris <wfh@chromium.org>
Reviewed-by: Ken Rockot <rockot@google.com>
Cr-Commit-Position: refs/heads/main@{#1108468}