Fixed the issue that slider sometimes isn't themed for dark/light mode
This issue is caused by sometimes the slider isn't repainted when the
dark/light mode changes. This is because GetColorProvider() is called
inside the OnPaint() method by slider itself, the views system won't
know about it when the theme changes and the slider needs a repaint.
So we manually call SchedulePaint() in the overridden OnThemeChanged()
to trigger the repaint.

Similar issue:
https://chromium-review.googlesource.com/c/chromium/src/+/4262693

Screen recording after the fix:
https://drive.google.com/file/d/1VLC5UfSCMhMU6h_k7Rhhdy6o2UIby9V0/view?usp=sharing

Bug: b/269636198
Change-Id: Ibf681077429da7c4b9960c258d11c87bed3a8e37
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4265599
Commit-Queue: Sylvie Liu <sylvieliu@chromium.org>
Reviewed-by: Andre Le <leandre@chromium.org>
Reviewed-by: Jiaming Cheng <jiamingc@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1108491}