[Trigger Attestation] Add metrics
Add 4 status metrics:
- Conversions.TriggerAttestation.GetHeadersStatus: Records the status of
the initial part of an attestation operation which attempts to retrieve
attestation headers to be added to a trigger registration request.

- Conversions.TriggerAttestation.ProcessAttestationStatus: Records the
status of the second part of an attestation operation which processes
the private state issuer's signature to generate an attestation token.

- Conversions.TriggerAttestation.ReportHasAttestation: Records the
status of the third and final part of an attestation operation which
attach an attestation token to an aggregatable report.

And one timing metric:
- Conversions.TriggerAttestation.Duration.{TimeSpan}.{Outcome}

Time taken to execute various parts of a Trigger Attestation operation:
[1] GetKeyCommitment = time taken to retrieve the issuer's key
    commitment.
[2] InitializeCryptographer = time take to initialize the BoringSSL
    cryptographer with the issuer's keys.
[3] BlindMessage = time taken by the cryptographer to blind the message.
[4] Issuer = time from the end of the outbound BlindMessage half to the
    beginning of the inbound UnblindSignature half.
[5] UnblindSignature = time taken to unblind the blind signature
    received from the issuer.
[6] Total = GetKeyCommitment + InitializeCryptographer + BlindMessage +
    UnblindSignature (only recorded for operations that see issuer's
    responses)

Each part of the operation can be a Failure or a Success. e.g. if we are
unable to retrieve key commitment for an issuer, it is a failure.

Bug: 1414404
Change-Id: I3789dd7d1a863ac7d0c18894c3171a16794a541e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4251197
Commit-Queue: Anthony Garant <anthonygarant@chromium.org>
Reviewed-by: John Delaney <johnidel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1108516}