Add temporary opaque colour for dogfood
For dogfood we are going to be testing with Jelly off, which means we need to correct the event list background colour. With Jelly on, it's now an opaque colour in both light and dark mode and with jelly off it falls back to a legacy colour.

Bug: b/270323641
Change-Id: I628d8ecf2136558c5fa162ee0c273df59fcf3c59
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4279708
Commit-Queue: Sam Cackett <samcackett@google.com>
Reviewed-by: Jiaming Cheng <jiamingc@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1108578}