[ImmersiveFullscreenTabs] Move tabs to the titlebar
This CL kicks off the work needed to move the tabs into the titlebar
while in immersive fullscreen. The codepath to enable this feature is
behind a feature flag: ImmersiveFullscreenTabs.

To enable this feature both the ImmersiveFullscreen and
ImmersiveFullscreenTabs feature flags must be used.
--enable-features="ImmersiveFullscreen,ImmersiveFullscreenTabs"

There are a number of TODOs to make the feature fully usable but this
CL provides a good base to work from.

https://drive.google.com/file/d/1qdbPtH5GYcbJwqSXooJbpDr4YCywfOir/view?usp=sharing&resourcekey=0-A7NxYaPGgg9zrZPcgOQm5w

Bug: 1414521
Change-Id: Ic6741fa1513c35337deed2409ab3ccc8d75aa08f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4250457
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Tom Burgin <bur@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1108609}