Settings: fix margin for settings section send feedback button
Previously, both a top and bottom margin was set on the feedback button,
however due to it being taller than the section header, this increased
the total height of the settings-section. This CL removes the bottom
margin, so that the header and feedback button keep the same relative
alignment, while reverting the settings-section to its original height.

Fixed: 1377051
Change-Id: I7a07794fbd72feb2c38fb71a844dc4f429301992
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4276663
Reviewed-by: John Lee <johntlee@chromium.org>
Commit-Queue: Charles Meng <charlesmeng@chromium.org>
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1108624}