[fuzzcoverage] Add a "hello world" builder for fuzzing coverage.
In order to create a fuzzing coverage dashboard, we need a new builder
that will (1) download corpora from Clusterfuzz, (2) build all Chromium
fuzzers, (3) run those fuzzers using those corpora as inputs, and (4)
output the result in the form of coverage information.

The linux-code-coverage builder already serves a similar function, but for unit
test coverage instead of fuzzing coverage. So, we add the new builder in
a similar location, and in a similar manner.

Note that the `chromium_fuzz_coverage` recipe referenced here is added
to the `chromium/tools/build/` repo in
https://chromium-review.googlesource.com/c/chromium/tools/build/+/4264408.

To reviewers: the goal of this CL is to have a builder, visible at
https://ci.chromium.org/p/chromium/builders/ci/ToTLinuxFuzzCoverage,
which runs periodically and outputs "success/fail" accordingly.  I
*believe* I've followed all the steps in
https://chromium.googlesource.com/chromium/src/+/HEAD/docs/infra/new_builder.md#Chromium-configuration,
but couldn't determine a way to "test" these changes pre-commit. So, any
help you could offer to ensure I've done this right would be very
appreciated.

Bug: 1356583
Change-Id: Idbf8651fbfd70db6f30ccfe0ba4a498f1ed8e9d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4265272
Reviewed-by: Prakhar Asthana <pasthana@google.com>
Commit-Queue: Julia Hansbrough <flowerhack@google.com>
Cr-Commit-Position: refs/heads/main@{#1108632}