Check return value of sql::MetaTable::SetVersionNumber/SetCompatibleVersionNumber
Without this, it is possible for a schema to be partially initialized,
e.g. the tables are changed but the version is not, which could lead to
unintended behavior, including crashes.

In the future, these methods will be marked [[nodiscard]] to help
prevent erroneous usage.

This CL was uploaded by git cl split.

R=tby@chromium.org

Bug: 1414092
Change-Id: I3e2e9db869cb03f674e376bc50baace8c5198202
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4280751
Auto-Submit: Andrew Paseltiner <apaseltiner@chromium.org>
Commit-Queue: Tony Yeoman <tby@chromium.org>
Reviewed-by: Tony Yeoman <tby@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1108643}