[QuickStart] Update endpoint info verification style
Change the verification style returned in the endpoint info to
DIGITS which indicates to the phone it should show a code for the
user to match.

We will eventually want to change this to OUT_OF_BAND when the QR
code UI is ready to be tested.

Bug: b/234655072
Test: Update existing test
Change-Id: Ic30875a5b3f439c5ee0b7d574b7891b410fadc92
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4279269
Commit-Queue: Curt Clemens <cclem@google.com>
Reviewed-by: Brittany Hartmire <bhartmire@google.com>
Reviewed-by: Michael Hansen <hansenmichael@google.com>
Cr-Commit-Position: refs/heads/main@{#1108648}